Change Account class fields from String to char[] for security
Collecting and storing sensitive information as String is less secure than char[] due to String's immutability.
Opens up the possibility of certain types of attacks that access memory. By using char[] we are able to clear the array
whenever we are done and shortens the window an attacker can exploit this vulnerability.

https://stackoverflow.com/questions/8881291/why-is-char-preferred-over-string-for-passwords